+++ scope-backend/migrations/20201201165015_journal.js
@@ -1,7 +1,7 @@
 
 exports.up = function (knex) {
   return knex.schema.createTable('journal', (table) => {
-    table.integer('id').notNullable().primary();
+    table.bigInteger('id').notNullable().primary();
     table.string('description').notNullable();
   });
 };
